+++ jaseci_core/jaseci/actions/standard/std.py
@@ -112,6 +112,18 @@ def actload_remote(url: str, meta):
 
 
 @jaseci_action()
+def actload_module(module: str, meta):
+    """
+    Load module actions to Jaseci
+    """
+    mast = master_from_meta(meta)
+    if(not mast.is_master(super_check=True, silent=True)):
+        meta['interp'].rt_error("Only super master can load actions.")
+        return False
+    return mast.actions_load_module(mod=module)['success']
+
+
+@jaseci_action()
 def destroy_global(name: str, meta):
     """Get utc date time for now in iso format"""
     mast = master_from_meta(meta)